lesson_5/requirements.txt
Flask==1.1.4
Werkzeug==0.16.1
python-dotenv==0.17.1
click==7.1.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.0.0
psycopg2-binary==2.8.6
WTForms==2.3.3
Flask-WTF==0.15.1
email-validator==1.1.2
